This commit should make it so that the pen can be drawn using a cribblePacket. In updateScreen(), there is a part where a cribblePacket and its parts are initialized by way of the current information.
This SHOULD work. I tried a different version where everything was based around cribblePacket. I went through every file and switched all instances that needed to be switched. I thought this way was simpler but it seems too easy...things are seldom this easy. This SHOULD work but I don't exactly know if it will work in all instances.